requirements.txt
Django
psycopg2
djangorestframework
coreapi
mock

